Improve Import Export configs

+++ app/src/main/java/com/drdisagree/iconify/utils/compiler/MonetCompiler.java
@@ -18,10 +18,10 @@ import java.io.IOException;
 public class MonetCompiler {
 
     private static final String TAG = MonetCompiler.class.getSimpleName();
-    private static boolean mEnable = false;
+    private static boolean mForce = false;
 
-    public static boolean buildOverlay(String[] resources, boolean enable) throws IOException {
-        mEnable = enable;
+    public static boolean buildOverlay(String[] resources, boolean force) throws IOException {
+        mForce = force;
 
         preExecute();
 
@@ -84,8 +84,8 @@ public class MonetCompiler {
         Shell.cmd("mkdir -p " + Resources.UNSIGNED_DIR).exec();
         Shell.cmd("mkdir -p " + Resources.SIGNED_DIR).exec();
 
-        // Disable the overlay in case it is already enabled
-        if (mEnable) {
+        if (mForce) {
+            // Disable the overlay in case it is already enabled
             OverlayUtil.disableOverlay("IconifyComponentME.overlay");
         }
     }
@@ -96,20 +96,23 @@ public class MonetCompiler {
             Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + Resources.OVERLAY_DIR + "/IconifyComponentME.apk").exec();
             RootUtil.setPermissions(644, Resources.OVERLAY_DIR + "/IconifyComponentME.apk");
 
-            // Move to files dir and install
-            Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
-            RootUtil.setPermissions(644, Resources.DATA_DIR + "/IconifyComponentME.apk");
-            Shell.cmd("pm install -r " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
-            Shell.cmd("rm -rf " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
+            if (mForce) {
+                // Move to files dir and install
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
+                RootUtil.setPermissions(644, Resources.DATA_DIR + "/IconifyComponentME.apk");
+                Shell.cmd("pm install -r " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
+                Shell.cmd("rm -rf " + Resources.DATA_DIR + "/IconifyComponentME.apk").exec();
 
-            SystemUtil.mountRW();
-            Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + "/system/product/overlay/IconifyComponentME.apk").exec();
-            RootUtil.setPermissions(644, "/system/product/overlay/IconifyComponentME.apk");
-            SystemUtil.mountRO();
+                // Move to system overlay dir
+                SystemUtil.mountRW();
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + Resources.SYSTEM_OVERLAY_DIR + "/IconifyComponentME.apk").exec();
+                RootUtil.setPermissions(644, "/system/product/overlay/IconifyComponentME.apk");
+                SystemUtil.mountRO();
 
-            // Enable the overlays
-            if (mEnable) {
+                // Enable the overlays
                 OverlayUtil.enableOverlays("IconifyComponentDM.overlay", "IconifyComponentME.overlay");
+            } else {
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponentME.apk " + Resources.BACKUP_DIR + "/IconifyComponentME.apk").exec();
             }
         }
