+++ app/src/main/java/com/drdisagree/iconify/utils/compiler/OnDemandCompiler.java
@@ -20,13 +20,13 @@ public class OnDemandCompiler {
     private static String mOverlayName = null;
     private static String mPackage = null;
     private static int mStyle = 0;
-    private static boolean mEnable = false;
+    private static boolean mForce = false;
 
-    public static boolean buildOverlay(String overlay_name, int style, String package_name, boolean enable) throws IOException {
+    public static boolean buildOverlay(String overlay_name, int style, String package_name, boolean force) throws IOException {
         mOverlayName = overlay_name;
         mPackage = package_name;
         mStyle = style;
-        mEnable = enable;
+        mForce = force;
 
         preExecute();
         moveOverlaysToCache();
@@ -83,8 +83,8 @@ public class OnDemandCompiler {
         Shell.cmd("mkdir -p " + Resources.SIGNED_DIR).exec();
         Shell.cmd("mkdir -p " + Resources.TEMP_CACHE_DIR + "/" + mPackage + "/").exec();
 
-        // Disable the overlay in case it is already enabled
-        if (mEnable) {
+        if (mForce) {
+            // Disable the overlay in case it is already enabled
             OverlayUtil.disableOverlay("IconifyComponent" + mOverlayName + ".overlay");
         }
     }
@@ -96,18 +96,22 @@ public class OnDemandCompiler {
             RootUtil.setPermissions(644, Resources.OVERLAY_DIR + "/IconifyComponent" + mOverlayName + ".apk");
 
             // Move to files dir and install
-            Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponent" + mOverlayName + ".apk " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
-            RootUtil.setPermissions(644, Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk");
-            Shell.cmd("pm install -r " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
-            Shell.cmd("rm -rf " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
+            if (mForce) {
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponent" + mOverlayName + ".apk " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
+                RootUtil.setPermissions(644, Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk");
+                Shell.cmd("pm install -r " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
+                Shell.cmd("rm -rf " + Resources.DATA_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
 
-            SystemUtil.mountRW();
-            Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponent" + mOverlayName + ".apk " + "/system/product/overlay/IconifyComponent" + mOverlayName + ".apk").exec();
-            RootUtil.setPermissions(644, "/system/product/overlay/IconifyComponent" + mOverlayName + ".apk");
-            SystemUtil.mountRO();
+                // Move to system overlay dir
+                SystemUtil.mountRW();
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponent" + mOverlayName + ".apk " + Resources.SYSTEM_OVERLAY_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
+                RootUtil.setPermissions(644, "/system/product/overlay/IconifyComponent" + mOverlayName + ".apk");
+                SystemUtil.mountRO();
 
-            if (mEnable) {
+                // Enable the overlay
                 OverlayUtil.enableOverlay("IconifyComponent" + mOverlayName + ".overlay");
+            } else {
+                Shell.cmd("cp -rf " + Resources.SIGNED_DIR + "/IconifyComponent" + mOverlayName + ".apk " + Resources.BACKUP_DIR + "/IconifyComponent" + mOverlayName + ".apk").exec();
             }
         }
